requirements.txt
python-telegram-bot==21.6
aiohttp==3.9.5
